Rename to shell package
