+++ src/main/resources/META-INF/services/javax.script.ScriptEngineFactory
@@ -1,4 +1,4 @@
 jsr223.nativeshell.bash.BashScriptEngineFactory
 jsr223.nativeshell.cmd.CmdScriptEngineFactory
 jsr223.nativeshell.executable.ExecutableScriptEngineFactory
-jsr223.nativeshell.shebang.ShellEngineFactory
+jsr223.nativeshell.shell.ShellEngineFactory
